moved the common stuff to util
* moved the common stuff to util, so that other apps can consume it
